Add separate configuration file
Configuration in a separate file works now. Was not working because
SERVER_NAME does not work on localhost. I need to read more on Flask. Oh,
the way the configuration was loaded is actually built into Flask and this
is not working properly. So far, the only configuration it can load seems
to be the SERVER_NAME, which is unnecessary on localhost of course.
Anything else, it can't find the key.