Refactor environment and command line parser
Script was not working from console because app path was not available to
PYTHONPATH environment variable, fixed that.

Currently the command line parser only supports running on a specific
host:port combination. Removed options for running on localhost by default
because that would not work if configurations are loaded from a file.

Also removed debug option from configuration file.